+++ tests/testthat/test-load_circuit_details.R
@@ -1,4 +1,5 @@
 test_that("load circuit details works", {
+  testthat::skip_if_offline("livetiming.formula1.com")
   skip_if_no_py()
   skip_if_no_ff1()