+++ NEWS.md
@@ -1,3 +1,7 @@
+# f1dataR dev
+* Removed `usethis` from Suggests (only needed when changing data)
+* Updated tests to avoid API failure for 2022 season & better skip-on-CRAN for internet resources
+
 # f1dataR 1.5.3
 * Enabled Sprint Qualifying "SQ" as a session type in line with FastF1 v3.3.5.