+++ NEWS.md
@@ -1,3 +1,4 @@
+
 # f1dataR 1.6.0.9000
 * Updates per FastF1 (python) updates at 3.4.0 #259
     * Soft deprecates FastF1 v < 3.4.0, hard deprecates FastF1 v < 3.1.0.
@@ -5,6 +6,8 @@
     * Added functions for look-ups of driver & team information for seasons/sessions. See `get_driver_abbreviation()`, `get_driver_name()`, `get_team_by_driver()`, `get_team_name()`, `get_drivers_by_team()`.
     * Removed `constructor_data` object from package vignettes, readme, examples, etc.
     * Expanded vignettes to improve graphics & demonstrate lookups.
+* Removed `usethis` from Suggests (only needed when changing data)
+* Updated tests to avoid API failure for 2022 season & better skip-on-CRAN for internet resources
 
 # f1dataR 1.5.3
 * Enabled Sprint Qualifying "SQ" as a session type in line with FastF1 v3.3.5.